+++ relay-metrics/src/aggregation.rs
@@ -17,7 +17,9 @@ use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tokio::time::Instant;
 
-use crate::statsd::{MetricCounters, MetricGauges, MetricHistograms, MetricSets, MetricTimers};
+use crate::statsd::{
+    metric_name_tag, MetricCounters, MetricGauges, MetricHistograms, MetricSets, MetricTimers,
+};
 use crate::{
     protocol, CounterType, DistributionType, GaugeType, Metric, MetricNamespace,
     MetricResourceIdentifier, MetricType, MetricValue, MetricsContainer, SetType,
@@ -1691,7 +1693,7 @@ impl AggregatorService {
             Entry::Occupied(mut entry) => {
                 relay_statsd::metric!(
                     counter(MetricCounters::MergeHit) += 1,
-                    metric_name = &entry.key().metric_name
+                    metric_name = metric_name_tag(&entry.key().metric_name),
                 );
                 let bucket_value = &mut entry.get_mut().value;
                 let cost_before = bucket_value.cost();
@@ -1702,11 +1704,11 @@ impl AggregatorService {
             Entry::Vacant(entry) => {
                 relay_statsd::metric!(
                     counter(MetricCounters::MergeMiss) += 1,
-                    metric_name = &entry.key().metric_name
+                    metric_name = metric_name_tag(&entry.key().metric_name),
                 );
                 relay_statsd::metric!(
                     set(MetricSets::UniqueBucketsCreated) = entry.key().hash64() as i64, // 2-complement
-                    metric_name = &entry.key().metric_name
+                    metric_name = metric_name_tag(&entry.key().metric_name),
                 );
 
                 let flush_at = self.config.get_flush_time(timestamp, project_key);